Update common widgets
